Solved bug #001
A check on the header-image.php was introduced to check if we're on the front page, or if we're on a post page. If this last is the case, we show up the header image, instead of trying to show the las image posted.
This commit is released as v. 0.2